[firewall_config] Rename iptables_config.* -> firewall_config.*
Update includes and CMakeLists.txt files.
