+++ src/platform/backends/qemu/qemu_virtual_machine_factory.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017-2020 Canonical, Ltd.
+ * Copyright (C) 2017-2021 Canonical, Ltd.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -18,7 +18,7 @@
 #define MULTIPASS_QEMU_VIRTUAL_MACHINE_FACTORY_H
 
 #include "dnsmasq_server.h"
-#include "iptables_config.h"
+#include "firewall_config.h"
 
 #include <multipass/path.h>
 #include <shared/base_virtual_machine_factory.h>
