+++ tests/qemu/test_firewall_config.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019 Canonical, Ltd.
+ * Copyright (C) 2019-2021 Canonical, Ltd.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -15,7 +15,7 @@
  *
  */
 
-#include <src/platform/backends/qemu/iptables_config.h>
+#include <src/platform/backends/qemu/firewall_config.h>
 
 #include "tests/mock_environment_helpers.h"
 #include "tests/mock_process_factory.h"
